+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Versions
 
+## 1.6.8
+
+- AutoFees: do not change inbound fee during pending HTLCs
+- AutoFees: bump Low Liq Rate of the channel's rule on failed HTLCs  
+
 ## 1.6.7
 
 - Fix AutoFees stop working bug